requirements.txt
playwright

google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2

gspread

beautifulsoup4

pandas

python-dotenv
